* split toolbox.py -> better overview, avoiding circular imports * add _replace_group_member_element_type in toolbox functions
