+++ pandapower/control/controller/const_control.py
@@ -4,7 +4,7 @@
 # and Energy System Technology (IEE), Kassel. All rights reserved.
 
 from pandapower.control.basic_controller import Controller
-from pandapower.toolbox import _detect_read_write_flag, write_to_net
+from pandapower.toolbox_elm_selection import _detect_read_write_flag, write_to_net
 
 try:
     import pandaplan.core.pplog as logging
