+++ pandapower/control/controller/trafo/ContinuousTapControl.py
@@ -5,7 +5,7 @@
 
 import numpy as np
 from pandapower.control.controller.trafo_control import TrafoController
-from pandapower.toolbox import read_from_net, write_to_net
+from pandapower.toolbox_elm_selection import read_from_net, write_to_net
 
 
 class ContinuousTapControl(TrafoController):
